rest: close idle connections to prevent EMFILE
If a client establishes HTTP connections to the REST API, the
connections may linger either due to a Connection: keep-alive, or
because of a malicious client performing a DoS attack.

These connections fill up the process open file quota, leading to
various other operations failing with EMFILE.  In my tests, the peer
grpc listener quit with a "too many open files" error, which lead to the
peer shutting down.

Limiting the duration of REST connections is no complete fix, but
is the least invasive fix against non-malicious REST clients keeping
their connections open.

Signed-off-by: Simon Schubert <sis@zurich.ibm.com>